requirements.txt
streamlit>=1.35.0
requests>=2.31.0
python-dotenv>=1.0.1
Pillow>=10.4.0
python-magic-bin>=0.4.14 